cudaMemset documentation says it takes an int and not a double. (See https://docs.nvidia.com/cuda/cuda-runtime-api/group__CUDART__MEMORY.html#group__CUDART__MEMORY_1gf7338650f7683c51ee26aadc6973c63a). Fixed this. All unit tests pass

+++ zero/alloc.c
@@ -133,7 +133,7 @@ gkyl_cu_memcpy(void *dst, void *src, size_t count, enum gkyl_cu_memcpy_kind kind
 }
 
 void
-gkyl_cu_memset(void *data, double val, size_t count)
+gkyl_cu_memset(void *data, int val, size_t count)
 {
   cudaMemset(data, val, count);
 }
@@ -176,7 +176,7 @@ gkyl_cu_memcpy(void *dst, void *src, size_t count, enum gkyl_cu_memcpy_kind kind
 }
 
 void
-gkyl_cu_memset(void *data, double val, size_t count)
+gkyl_cu_memset(void *data, int val, size_t count)
 {
   assert(false);  
 }
