+++ zero/gkyl_alloc.h
@@ -62,4 +62,4 @@ void gkyl_cu_free_host(void *ptr);
 void gkyl_cu_memcpy(void *dst, void *src, size_t count, enum gkyl_cu_memcpy_kind kind);
 
 /** Set memory on device */
-void gkyl_cu_memset(void *data, double val, size_t count);
+void gkyl_cu_memset(void *data, int val, size_t count);
